+++ CHANGES.md
@@ -6,6 +6,7 @@
 -   Make pagination on mobile responsive.
 -   Fixed a facet overflow issue on desktop
 -   Allow user apply default date in filter
+-   Fixed Not found redirect working differently in different browsers
 
 ## 0.0.45